Make Tracker and ForwardDiff optional dependencies
