+++ src/Bijectors.jl
@@ -1,9 +1,7 @@
 module Bijectors
 
-using Reexport
+using Reexport, Requires
 @reexport using Distributions
-using ForwardDiff
-using Tracker
 using StatsFuns
 using LinearAlgebra
 using MappedArrays
@@ -45,11 +43,8 @@ export  TransformDistribution,
 
 const DEBUG = Bool(parse(Int, get(ENV, "DEBUG_BIJECTORS", "0")))
 
-# Workaround for eps(::ForwardDiff.Dual)
 _eps(::Type{T}) where {T} = T(eps(T))
 _eps(::Type{Real}) = eps(Float64)
-_eps(::Type{<:ForwardDiff.Dual{<:Any, Real}}) = _eps(Real)
-_eps(::Type{<:Tracker.TrackedReal{T}}) where {T} = eps(T)
 
 #=
   NOTE: Codes below are adapted from
@@ -532,5 +527,11 @@ function logpdf_with_trans(
 end
 
 include("interface.jl")
+
+# optional dependencies
+function __init__()
+    @require ForwardDiff="f6369f11-7733-5829-9624-2563aa707210" include("forwarddiff.jl")
+    @require Tracker="9f7883ad-71c0-57eb-9f7f-b5c9e6d3789c" include("tracker.jl")
+end
 
 end # module
