+++ src/bijectors/adbijector.jl
@@ -4,34 +4,15 @@ implement `jacobian` and, by impliciation, `logabsdetjac`.
 """
 abstract type ADBijector{AD, N} <: Bijector{N} end
 
-# AD implementations
-function jacobian(b::Union{B, Inversed{B}}, x::Real) where {B<:ADBijector{<:ForwardDiffAD}}
-    return ForwardDiff.derivative(b, x)
-end
-function jacobian(
-    b::Union{B, Inversed{B}},
-    x::AbstractVector{<:Real}
-) where {B<:ADBijector{<:ForwardDiffAD}}
-    return ForwardDiff.jacobian(b, x)
-end
-
-function jacobian(b::Union{B, Inversed{B}}, x::Real) where {B<:ADBijector{<:TrackerAD}}
-    return Tracker.data(Tracker.gradient(b, x)[1])
-end
-function jacobian(
-    b::Union{B, Inversed{B}},
-    x::AbstractVector{<:Real}
-) where {B<:ADBijector{<:TrackerAD}}
-    # We extract `data` so that we don't return a `Tracked` type
-    return Tracker.data(Tracker.jacobian(b, x))
-end
-
 struct SingularJacobianException{B<:Bijector} <: Exception
     b::B
 end
 function Base.showerror(io::IO, e::SingularJacobianException)
     print(io, "jacobian of $(e.b) is singular")
 end
+
+# concrete implementations with optional dependencies ForwardDiff and Tracker
+function jacobian end
 
 # TODO: allow batch-computation, especially for univariate case?
 "Computes the absolute determinant of the Jacobian of the inverse-transformation."
